Move function show to a specific golang file
